Create lists2dict
A python fn that creates a dictionary when provided with two lists. The lists are zipped using the zip() fn to create a list of tuples which is then converted to a dictionary with the first element of each tuple being the key and the second being the value of the pair.